backend/requirements.txt
fitz==0.0.1.dev2
keras_ocr==0.9.2
numpy==1.26.1
opencv_contrib_python==4.8.1.78
Pillow==10.1.0
